test: Move git-specific test over to git test file
